\name{landscape.amova.pairwise}
\alias{landscape.amova.pairwise}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ calculates pairwise phi-ST for a landscape}
\description{
  pairwise \eqn{\phi_{ST}} calculator.  Kind of slow. use \code{landscape.sample}
}
\usage{
landscape.amova.pairwise(rland)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{rland}{landscape object}
}

\seealso{ landscape.amova, landscape.amova.locus}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc}

