\name{landscape.make.genind}
\alias{landscape.make.genind}
\title{Converts genetic marker data in a landscape into a the genind
  format from adegenet}

\description{
Makes a genind object from a landscape.  Populations can also be given
names with an optional parameter.
}
\usage{
   landscape.make.genind(Rland)
}
\arguments{
  \item{Rland}{the Rmetasim landscape object}
  
}
\value{
  an adegenet 'genind' object with optional population names
}
\seealso{
landscape.make.genpop
}

\keyword{misc}
