\name{landscape.make.genpop}
\alias{landscape.make.genpop}
\title{Converts genetic marker data in a landscape into a the genpop
  format from adegenet}

\description{
Makes a genpop object from a landscape.
}
\usage{
   landscape.make.genpop(Rland)
}
\arguments{
  \item{Rland}{the Rmetasim landscape object}
}
\value{
  an adegenet 'genpop' object.
}
\seealso{
landscape.make.genind
}

\keyword{misc}
